feat: ReactNews 자체적으로 hover 및 페칭을 처리하도록 개선
