+++ src/components/AppHeader/RecentNewsRoll.ts
@@ -1,18 +1,27 @@
 import { A, Div, Span } from "../../libs/Elements";
-import { News } from "../../models/News";
 import { getMediaById } from "../../remotes/getMediaById";
 import typoStyles from "@/styles/typo.module.css";
-import styles from "./RecentNews.module.css";
+import styles from "./RecentNewsRoll.module.css";
+import { useState, useEffect } from "@/libs/createApp";
+import { getRecentNews } from "@/remotes/getRecentNewsList";
 
-export const RecentNews = ({
-  mediaId,
-  news,
-  setIsHovering,
-}: {
-  mediaId: number;
-  news: News;
-  setIsHovering: (isHovering: boolean) => void;
-}) => {
+export const RecentNewsRoll = ({ differ }: { differ: number }) => {
+  const [getIsHovering, setIsHovering] = useState(false);
+  const [getFrom, setFrom] = useState(0);
+
+  const { news, mediaId } = getRecentNews({
+    from: getFrom() * 2 + differ,
+    limit: 2,
+  });
+  useEffect(() => {
+    const interval = setInterval(() => {
+      if (getIsHovering()) return;
+      setFrom(getFrom() + 1);
+    }, 5000);
+    return () => {
+      clearInterval(interval);
+    };
+  });
   const media = getMediaById(mediaId);
 
   return Div({
